tests/wpt/web-platform-tests/offscreen-canvas/text/2d.text.measure.width.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.measure.width.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>2d.text.measure.width.basic</h1>
<p class="desc">The width of character is same as font used for OffscreenCanvas</p>


<script>
var t = async_test("The width of character is same as font used for OffscreenCanvas");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var f = new FontFace("CanvasTest", "/fonts/CanvasTest.ttf");
let fonts = (self.fonts ? self.fonts : document.fonts);
fonts.add(f);
fonts.ready.then(() => {
    return new Promise(function(resolve) { step_timeout(resolve, 500); });
}).then(function() {
    ctx.font = '50px CanvasTest';
    _assertSame(ctx.measureText('A').width, 50, "ctx.measureText('A').width", "50");
    _assertSame(ctx.measureText('AA').width, 100, "ctx.measureText('AA').width", "100");
    _assertSame(ctx.measureText('ABCD').width, 200, "ctx.measureText('ABCD').width", "200");

    ctx.font = '100px CanvasTest';
    _assertSame(ctx.measureText('A').width, 100, "ctx.measureText('A').width", "100");
}).then(t_pass, t_fail);

});
</script>
