src/third_party/blink/web_tests/platform/mac/external/wpt/offscreen-canvas/text/2d.text.measure.width.basic-expected.txt

This is a testharness.js-based test.
FAIL The width of character is same as font used for OffscreenCanvas assert_equals: ctx.measureText('A').width === 50 (got 36.1083984375[number], expected 50[number]) expected 50 but got 36.1083984375
Harness: the test ran to completion.
